requirements.txt
amqp==5.1.1
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
blis==0.7.7
cached-property==1.5.2
catalogue==1.0.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.3
cryptography==37.0.2
cymem==2.0.6
defusedxml==0.7.1
Django==3.2
django-currentuser==0.5.3
django-debug-toolbar==3.2.4
django-embed-video==1.4.4
django-environ==0.8.1
django-filter==21.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
Faker==8.10.3
greenlet==1.1.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
kombu==5.2.4
MarkupSafe==2.1.1
mimesis==4.1.3
model-bakery==1.5.0
more-itertools==8.13.0
murmurhash==1.0.7
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
openapi-codec==1.3.2
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==21.3
Pillow==9.1.0
pipdeptree==2.2.1
plac==1.1.3
pluggy==0.13.1
poetry-version==0.1.5
preshed==3.0.6
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==3.5.3
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.15.100
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
spacy==2.3.7
SQLAlchemy==1.4.36
sqlparse==0.4.2
srsly==1.0.5
text-unidecode==1.3
thinc==7.4.5
toml==0.10.2
tomlkit==0.5.11
tqdm==4.64.0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wasabi==0.9.1
wcwidth==0.2.5
gunicorn==19.9.0